% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{binarize}
\alias{binarize}
\title{Binarize numeric vector}
\usage{
binarize(vec)
}
\arguments{
\item{vec}{numeric}
}
\value{
vec numeric
}
\description{
Binarize a numeric vector using kmeans clustering (k=2)
}
\examples{
binarize(rnorm(100))
}
