% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{min_max_normalize}
\alias{min_max_normalize}
\title{Min max normalization}
\usage{
min_max_normalize(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Min max normalization
}
\examples{
min_max_normalize(rnorm(10))
}
