% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{open_file_mac}
\alias{open_file_mac}
\title{Open file}
\usage{
open_file_mac(file)
}
\arguments{
\item{file}{character}
}
\value{
none
}
\description{
Open a file in mac OS.
}
\examples{
write("test", "test.txt")
open_file_mac("test.txt")
}
