% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{scale_rows}
\alias{scale_rows}
\title{Scale rows}
\usage{
scale_rows(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Scales the rows of a numeric matrix.
}
\examples{
x <- rnorm(100, 10, 10)
scale_rows(x)
}
