% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{sigmoid_transform}
\alias{sigmoid_transform}
\title{Sigmoid transform}
\usage{
sigmoid_transform(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Sigmoid transform
}
\examples{
sigmoid_transform(rnorm(10))
}
