% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{str_extract_pos}
\alias{str_extract_pos}
\title{Extract character part}
\usage{
str_extract_pos(charvec, patt, inx, ...)
}
\arguments{
\item{charvec}{character}

\item{patt}{character}

\item{inx}{integer}

\item{...}{arguments for strsplit}
}
\value{
character
}
\description{
Extract the subset of characters at position inx after splitting at pattern patt.
}
\examples{
str_extract_pos("split_this_at_patt", "_", 2)
}
