% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{vector_to_colors}
\alias{vector_to_colors}
\title{Map character vector to colors}
\usage{
vector_to_colors(vec, cpall = "npg")
}
\arguments{
\item{vec}{character vector}

\item{cpall}{character}
}
\value{
character vector
}
\description{
Map character vector to colors
}
\examples{
vector_to_colors(c("Yes", "Yes", "No", "No"))
}
